requirements.txt
# requirements
streamlit
pandas
numpy
matplotlib
python-dotenv
openmeteo-requests
requests-cache 
retry-requests